+++ app/src/main/java/com/mobile/todo/adapter/ToDoAdapter.kt
@@ -1,5 +1,7 @@
 package com.mobile.todo.adapter
 
+import android.app.AlertDialog
+import android.app.Dialog
 import android.content.Intent
 import android.graphics.Paint
 import android.view.LayoutInflater
@@ -8,6 +10,7 @@ import android.view.ViewGroup
 import android.widget.CheckBox
 import android.widget.ImageView
 import android.widget.TextView
+import android.widget.Toast
 import androidx.core.content.ContextCompat.startActivity
 import androidx.recyclerview.widget.RecyclerView
 import com.mobile.todo.EditTodoHabit
@@ -16,6 +19,7 @@ import com.mobile.todo.database.AppDatabase
 import com.mobile.todo.database.dataset.Search
 import com.mobile.todo.database.dataset.Tag
 import com.mobile.todo.database.dataset.ToDo
+import com.mobile.todo.utils.Permission
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 
@@ -49,14 +53,26 @@ class ToDoAdapter(private var itemList: MutableList<ToDo>, private var searchLis
         }
 
         holder.delete.setOnClickListener {
-            GlobalScope.launch {
-                AppDatabase.getDatabase(context.context).toDoDao().deleteToDo(item)
-            }
-            val currentPosition = itemList.indexOf(item)
-            if (currentPosition != -1) {
-                itemList.removeAt(currentPosition)
-                notifyItemRemoved(currentPosition)
-            }
+            AlertDialog.Builder(context.context)
+                .setTitle("Delete to do")
+                .setMessage("Do you want to delete this to do?")
+                .setPositiveButton("OK") { dialog, _ ->
+                    // Remove to do from database
+                    GlobalScope.launch {
+                        AppDatabase.getDatabase(context.context).toDoDao().deleteToDo(item)
+                    }
+                    val currentPosition = itemList.indexOf(item)
+                    if (currentPosition != -1) {
+                        itemList.removeAt(currentPosition)
+                        notifyItemRemoved(currentPosition)
+                    }
+                    dialog.dismiss()
+                }
+                .setNegativeButton("Cancel") { dialog, _ ->
+                    dialog.dismiss()
+                }
+                .create()
+                .show()
         }
 
         holder.textView.setOnClickListener {
